+++ R/drapery.R
@@ -410,8 +410,7 @@ drapery <- function(x, type = "zvalue", layout = "grayscale",
   ##
   if (missing(xlab)) {
     xlab <- ""
-    xlab <- xlab(x$sm, backtransf = backtransf,
-                   efficacy = inherits(x, c("metabin", "metainc")))
+    xlab <- xlab(x$sm, backtransf = backtransf)
   }
   else
     chkchar(xlab, length = 1)