+++ R/funnel.R
@@ -435,8 +435,7 @@ funnel.meta <- function(x,
     else if (sm == "PRAW")
       xlab <- "Proportion"
     else
-      xlab <- xlab(sm, FALSE,
-                   efficacy = inherits(x, c("metabin", "metainc")))
+      xlab <- xlab(sm, FALSE)
   ##
   if (is.null(xlim) & !is.null(level) &
       (yaxis == "se" |